Fixed issue 37 - the cron process is now optimized to load any file available if the data is not complete in the DB
Fixed issue 37:
https://github.com/stawen/okovision/issues/37

Now all files that are available in the boiler will be downloaded if
needed. If for some reason, Okofen updates their software and let us
have access to more than 4 days of history, the code will still work and
be able to upload as many files as available.

Changed also the behaviour of the cron process: now we don't care when
the cron is done, we will automatically build the syntheses that are
missing for the last days (and systematically rebuild any synthese if we
happen to reload the data file)